big refactor to begin unifying components
