+++ ohmg/frontend/svelte/src/helpers/ol-styles.js
