+++ ohmg/frontend/svelte/src/helpers/utils.js
@@ -256,3 +256,10 @@ export function removeRotateCenter (layer) {
 export const iconProps = {
 	weight: 'bold',
 }
+
+export function setMapExtent(map, extent4326) {
+	if (map) {
+		const extent3857 = transformExtent(extent4326, "EPSG:4326", "EPSG:3857");
+		map.getView().fit(extent3857);
+	}
+}
