+++ ohmg/frontend/svelte/src/lib/components/buttons/PaginationButtons.svelte
@@ -1,6 +1,6 @@
 <script>
     import IconContext from 'phosphor-svelte/lib/IconContext';
-    import { iconProps } from "../../../js/utils"
+    import { iconProps } from "@helpers/utils"
 
     import CaretDoubleLeft from 'phosphor-svelte/lib/CaretDoubleLeft'
     import CaretDoubleRight from 'phosphor-svelte/lib/CaretDoubleRight'
