+++ ohmg/frontend/svelte/src/lib/components/interfaces/modals/LegendModal.svelte
@@ -1,0 +1,11 @@
+<script>
+    import Modal from '@components/modals/Base.svelte';
+
+    export let id;
+    export let legendUrl;
+    export let legendAlt;
+</script>
+
+<Modal id={id}>
+    <img src={legendUrl} alt={legendAlt} />
+</Modal>
