+++ ohmg/frontend/svelte/src/resource.js
@@ -1,5 +1,5 @@
-import '../css/interface.css';
-import Resource from '../lib/Resource.svelte';
+import '@src/css/interface.css';
+import Resource from '@src/lib/Resource.svelte';
 
 const app = new Resource({
 	target: document.getElementById("resource-target"),
