+++ ohmg/frontend/svelte/src/viewer.js
@@ -1,6 +1,6 @@
-import Viewer from '../lib/Viewer.svelte';
+import PlaceViewer from '@components/interfaces/PlaceViewer.svelte';
 
-const viewer = new Viewer({
+const viewer = new PlaceViewer({
 	target: document.getElementById("viewer-target"),
 	props: JSON.parse(document.getElementById("viewer-props").textContent),
 });
