+++ ohmg/frontend/svelte/src/volume.js
@@ -1,8 +1,8 @@
-import '../css/shared.css';
-import '../css/interface.css';
+import '@src/css/shared.css';
+import '@src/css/interface.css';
 
-import Volume from '../lib/Volume.svelte';
-import '../css/ol-overrides.css'
+import Volume from '@src/lib/Volume.svelte';
+import '@src/css/ol-overrides.css'
 
 const volume = new Volume({
 	target: document.getElementById("volume-target"),
